packages/api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
supabase
pyyaml
openai
httpx
# zep-python  # 将在Memory Engine集成时添加